Initial implementation of skeleton
